testing deploy

+++ scripts/application_stop.sh
@@ -1,4 +1,7 @@
 echo "stopping any existing node servers"
-sudo systemctl stop EmailApi.service
-sudo systemctl disable EmailApi.service
-sudo rm -f /etc/systemd/system/EmailApi.service
+STATUS="$(systemctl is-active EmailApi.service)"
+if [ "${STATUS}" = "active" ]; then
+    sudo systemctl stop EmailApi.service
+    sudo systemctl disable EmailApi.service
+    sudo rm -f /etc/systemd/system/EmailApi.serviceecho "Execute your tasks ....."
+fi
